+++ lib/get_rx/src/rx_types/rx_types.dart
@@ -31,6 +31,7 @@ part "rx_core/rx_impl.dart";
 part "rx_core/rx_interface.dart";
 part "rx_core/rx_num.dart";
 part "rx_core/rx_string.dart";
+part "rx_core/json_utils.dart";
 part "rx_iterables/rx_list.dart";
 part "rx_iterables/rx_map.dart";
 part "rx_iterables/rx_set.dart";